Fix menu pane bug when resizing window
This fixes a bug where the user can shrink the window to enable mobile
mode and bring the menu up. The menu now has a shadown box and the main
content is darker so that the focus is on the menu. The bug is that at
this moment (menu is open), if the window is enlarged back to desktop
mode, then the shadow box and the dark filter above the content don't
disappear.